Move py_version and processor fixture to utils.py
