+++ test/integration/local/test_horovod.py
@@ -19,17 +19,9 @@ import tarfile
 import pytest
 from sagemaker.tensorflow import TensorFlow
 
+from test.integration.sagemaker.utils import processor, py_version  # noqa: F401
+
 RESOURCE_PATH = os.path.join(os.path.dirname(__file__), '..', '..', 'resources')
-
-
-@pytest.fixture(params=os.environ['TEST_PY_VERSIONS'].split(','))
-def py_version(request):
-    return request.param
-
-
-@pytest.fixture(params=os.environ['TEST_PROCESSORS'].split(','))
-def processor(request):
-    return request.param
 
 
 @pytest.mark.skip_gpu
