+++ test/integration/sagemaker/test_horovod.py
@@ -14,23 +14,12 @@ from __future__ import absolute_import
 
 import os
 
-import pytest
 import sagemaker
 from sagemaker.tensorflow import TensorFlow
 
-from utils import unique_name_from_base
+from utils import processor, py_version, unique_name_from_base  # noqa: F401
 
 RESOURCE_PATH = os.path.join(os.path.dirname(__file__), '..', '..', 'resources')
-
-
-@pytest.fixture(params=os.environ['TEST_PY_VERSIONS'].split(','))
-def py_version(request):
-    return request.param
-
-
-@pytest.fixture(params=os.environ['TEST_PROCESSORS'].split(','))
-def processor(request):
-    return request.param
 
 
 def test_distributed_training_horovod(sagemaker_session,
