+++ test/integration/sagemaker/test_tuning_model_dir.py
@@ -14,21 +14,10 @@ from __future__ import absolute_import
 
 import os
 
-import pytest
 from sagemaker.tensorflow import TensorFlow
 from sagemaker.tuner import HyperparameterTuner, IntegerParameter
 
-from utils import unique_name_from_base
-
-
-@pytest.fixture(params=os.environ['TEST_PY_VERSIONS'].split(','))
-def py_version(request):
-    return request.param
-
-
-@pytest.fixture(params=os.environ['TEST_PROCESSORS'].split(','))
-def processor(request):
-    return request.param
+from utils import processor, py_version, unique_name_from_base  # noqa: F401
 
 
 def test_model_dir_with_training_job_name(sagemaker_session, ecr_image, instance_type, framework_version):
